requirements.txt
qrcode==7.3.1
Pillow==9.5.0
numpy==1.24.3
matplotlib>=3.6.0
Flask==2.0.1
Flask-Cors==3.0.10
pytest==7.4.0
pytest-cov==4.1.0
gunicorn==20.1.0
